Better error messages and exercices testing

+++ serverAPI/CodeTester.php
@@ -38,11 +38,17 @@ class CodeTester
 
         $containerName = "sandbox" . uniqid();
 
-        $dockerCmd = "docker run --rm --name $containerName -v $uploadPath:/tmp/code.cpp:ro -v $testFile:/tmp/test.cpp:ro sandbox bash -c 'export LANG=en_US.UTF-8; g++ -std=c++20 /tmp/code.cpp -o /tmp/code_exec && g++ -std=c++20 /tmp/test.cpp -o /tmp/test_exec -lgtest -lgtest_main -pthread && /tmp/test_exec --gtest_output=json:res.json > /dev/null 2>&1; cat res.json'";
+        $dockerCmd = "docker run --rm --name $containerName -v $uploadPath:/tmp/code.cpp:ro -v $testFile:/tmp/test.cpp:ro sandbox bash -c '
+export LANG=en_US.UTF-8;
+if ! g++ -std=c++20 /tmp/code.cpp -o /tmp/code_exec 2> /tmp/res.json; then cat /dev/null; fi;
+if ! g++ -std=c++20 /tmp/test.cpp -o /tmp/test_exec -lgtest -lgtest_main -pthread 2>> /tmp/res.json; then cat /tmp/res.json; exit; fi;
+/tmp/test_exec --gtest_output=json:/tmp/res.json > /dev/null 2>&1;
+cat /tmp/res.json'";
+
 
         $temp = "docker run -it --rm --name $containerName -v $uploadPath:/tmp/code.cpp:ro -v $testFile:/tmp/test.cpp:ro sandbox bash";
 
-        $temp2 = "g++ -std=c++20 /tmp/code.cpp -o /tmp/code_exec && g++ -std=c++20 /tmp/test.cpp -o /tmp/test_exec -lgtest -lgtest_main -pthread && /tmp/test_exec --gtest_output=json:res.json > /dev/null 2>&1; cat res.json";
+        $temp2 = "g++ -std=c++20 /tmp/code.cpp -o /tmp/code_exec 2>/dev/null ; g++ -std=c++20 /tmp/test.cpp -o /tmp/test_exec -lgtest -lgtest_main -pthread && /tmp/test_exec --gtest_output=json:res.json > /dev/null 2>&1; cat res.json";
 
         $testOutput = shell_exec($dockerCmd);
